renaming ssnmf to tsnmf
