+++ tsnmf/tests/test_smoothNMF.py
@@ -1,6 +1,6 @@
 import numpy.testing as nt
 from scipy.io import loadmat
-import ssnmf
+import tsnmf
 import numpy.linalg as LA
 
 
@@ -14,7 +14,7 @@ data_path = op.join(ssnmf.__path__[0],'data')
 def test_all_zero_one_iteration():
     # testing no sparsity, no smoothnes, 1 iteration
     output = loadmat(op.join(data_path,'output.mat'))
-    model = ssnmf.smoothNMF(n_components=5, max_iter=1, betaW=0, betaH=0)
+    model = tsnmf.smoothNMF(n_components=5, max_iter=1, betaW=0, betaH=0)
     model.fit(output['V'], W=output['init_W'], H=output['init_H'], init='custom')
 
     print(LA.norm(output['V'] - (model.W@model.H)))
@@ -31,7 +31,7 @@ def test_all_zero_one_iteration():
 def test_all_zero_200_iterations():
     # testing no sparsity, no smoothness, 200 iterations
     output = loadmat(op.join(data_path,'output.mat'))
-    model = ssnmf.smoothNMF(n_components=5, max_iter=200, betaW=0, betaH=0)
+    model = tsnmf.smoothNMF(n_components=5, max_iter=200, betaW=0, betaH=0)
     model.fit(output['V'], W=output['init_W'], H=output['init_H'], init='custom')
     nt.assert_almost_equal(model.cost[-1],3636.162716116)
 
@@ -39,7 +39,7 @@ def test_all_zero_200_iterations():
 def sparse_one_iteration():
     # testing sparsity, 1 iteration
     output = loadmat(op.join(data_path,'output.mat'))
-    model = ssnmf.smoothNMF(n_compoments=5, max_iter=1, sparsity=1, smoothness=0, betaW=0, betaH=0)
+    model = tsnmf.smoothNMF(n_compoments=5, max_iter=1, sparsity=1, smoothness=0, betaW=0, betaH=0)
     model.fit(output['V'], W=output['init_W'], H=output['init_H'], init='custom')
     nt.assert_almost_equal(model.cost[-1],4750.738752595)
 
@@ -47,7 +47,7 @@ def sparse_one_iteration():
 def test_smooth_one_iteration():
     # testing smoothness, 1 iteration
     output = loadmat(op.join(data_path,'output.mat'))
-    model = ssnmf.smoothNMF(n_components=5, max_iter=1, sparsity=0, smoothness=1, betaW=0.0, betaH=0.0)
+    model = tsnmf.smoothNMF(n_components=5, max_iter=1, sparsity=0, smoothness=1, betaW=0.0, betaH=0.0)
     model.fit(output['V'], W=output['init_W'], H=output['init_H'], init='custom')
 
     import numpy.linalg as LA
@@ -61,7 +61,7 @@ def test_smooth_one_iteration():
 def test_smooth_and_parse_one_iterations():
     # testing sparsity and smoothness, 1 iteration
     output = loadmat(op.join(data_path,'output.mat'))
-    model = ssnmf.smoothNMF(n_components=5, max_iter=1, sparsity=1, smoothness=1, betaW=0, betaH=0)
+    model = tsnmf.smoothNMF(n_components=5, max_iter=1, sparsity=1, smoothness=1, betaW=0, betaH=0)
     model.fit(output['V'], W=output['init_W'], H=output['init_H'], init='custom')
     nt.assert_almost_equal(model.cost[-1],6715.167611171)
 
@@ -69,13 +69,13 @@ def test_smooth_and_parse_one_iterations():
 def test_smooth_and_parse_200_iterations():
     # testing sparsity and smoothness, 200 iterations
     output = loadmat(op.join(data_path,'output.mat'))
-    model = ssnmf.smoothNMF(n_components=5, max_iter=200, sparsity=1, smoothness=1, betaW=0.0, betaH=0.0)
+    model = tsnmf.smoothNMF(n_components=5, max_iter=200, sparsity=1, smoothness=1, betaW=0.0, betaH=0.0)
     model.fit(output['V'], W=output['init_W'], H=output['init_H'], init='custom')
     nt.assert_almost_equal(model.cost[-1],3909.6946, decimal=4)
 
 def test_smooth_and_sparse_200_iterations_betas():
     # testing sparsity and smoothness, 200 iterations
     output = loadmat(op.join(data_path,'output.mat'))
-    model = ssnmf.smoothNMF(n_components=5, max_iter=200, sparsity=1, smoothness=1, betaW=0.1, betaH=0.1)
+    model = tsnmf.smoothNMF(n_components=5, max_iter=200, sparsity=1, smoothness=1, betaW=0.1, betaH=0.1)
     model.fit(output['V'], W=output['init_W'], H=output['init_H'], init='custom')
     nt.assert_almost_equal(model.cost[-1],3893.69665, decimal=4)
